Extract cryptographic functions
