+++ lib/r2d2.rb
@@ -1,3 +1,7 @@
 require "json"
+require 'openssl'
+require 'base64'
+require 'hkdf'
 
+require_relative "r2d2/util"
 require_relative "r2d2/payment_token"
